修复大屏导出失败问题

+++ simo-leaderview/src/main/java/com/uxsino/leaderview/service/HomeTemplateImgService.java
@@ -112,7 +112,7 @@ public class HomeTemplateImgService {
         return imgDao.findOne(id);
     }
 
-    public Integer getMaxId() {
+    public Long getMaxId() {
         return imgDao.getMaxId();
     }
